Implement buffer

+++ app/src/main/java/io/jim/tesserapp/graphics/Vertex.kt
@@ -25,7 +25,7 @@ data class Vertex(
 ) : Iterable<Float> {
 
     private val floats = listOf(
-            position.x.toFloat(), position.y.toFloat(), position.z.toFloat(),
+            position.x, position.y, position.z,
             color.red, color.green, color.blue,
             modelIndex.toFloat()
     )
